+++ django_tables2/tables.py
@@ -226,7 +226,8 @@ class TableBase(object):
 
         self.exclude = exclude or self._meta.exclude
         self.sequence = sequence
-        self.data = TableData.from_data(data=data, table=self)
+        self.data = TableData.from_data(data=data)
+        self.data.table = self
         if default is None:
             default = self._meta.default
         self.default = default